+++ aerosense/__init__.py
@@ -13,8 +13,8 @@ Usage:
 from typing import List
 
 # --- Versioning ---
-__version__: str = "v3.16.10"
-__release__: str = "2026-03-22"
+__version__: str = "v4.0.0"
+__release__: str = "2026-03-23"
 
 # --- Component Exports ---
 from .core.controller import Controller